close #173 split SceneGame.cpp file
